Imports in mitiq_qiskit. Removes unused seed for random_identity_circuit + renames function. Moves `measure` from utils to the test file.
